Add files via upload
This project aims to predict if the pulsar is a pulsar or not. It uses machine learning to predict the pulsar. The dataset has been used in this project was downloaded from kaggle. 
I have used SVM classifier model with different parameters by doing hyperparameter tuning.

All the model accuracies are above 90%.